Bug report: global vue mixin
A mixin registered by `Vue.mixin` is global, which changes Vue.js everywhere in the project. For example you can delete `import Generic` and `mixins: [Generic]` and everything is still working. This is not proper to do in a library, will you look into the code and find every occasions where Generic is used so that we can fix this?